refactors layout to be more readable
