+++ lib/sources/presentation/layouts/converter_page/converter_page_app_bar.dart
@@ -1,12 +1,15 @@
-import 'package:colored/sources/presentation/layouts/names_list/color_names_app_bar.dart';
+import 'package:colored/sources/domain/view_models/names_list/names_list_data.dart';
+import 'package:colored/sources/presentation/layouts/color_suggestions/color_suggestions_layout.dart';
+import 'package:colored/sources/presentation/layouts/color_search/color_search_layout.dart';
 import 'package:colored/sources/presentation/layouts/naming/naming_layout.dart';
 import 'package:colored/sources/presentation/widgets/animations/cross_switchers/cross_animated_switcher.dart';
 import 'package:colored/sources/presentation/widgets/animations/cross_switchers/cross_switcher_state.dart';
 
 import 'package:flutter/material.dart';
 
-class ConverterAppBar extends StatelessWidget implements PreferredSizeWidget {
-  const ConverterAppBar({
+class ConverterPageAppBar extends StatelessWidget
+    implements PreferredSizeWidget {
+  const ConverterPageAppBar({
     @required this.searchingState,
     this.onSearchStateChange,
     Key key,
@@ -19,7 +22,12 @@ class ConverterAppBar extends StatelessWidget implements PreferredSizeWidget {
   Widget build(BuildContext context) => CrossAnimatedSwitcher(
         state: searchingState,
         firstChild: NamingLayout(onSearchPressed: onSearchStateChange),
-        secondChild: ColorNamesAppBar(onBackPressed: onSearchStateChange),
+        secondChild: ColorSearchLayout(
+          flexibleSpaceChild: ColorSuggestionsLayout(
+            onSuggestionSelected: NamesListData.of(context).onSearchChanged,
+          ),
+          onBackPressed: onSearchStateChange,
+        ),
       );
 
   @override
